feat(interface-types) Implement the `call` executable instruction.
The patch requires to implement the `wasm::TypedIndex`,
`wasm::LocalImportIndex`, and the `wasm::LocalImport` traits.

+++ lib/interface-types/src/decoders/binary.rs
@@ -138,7 +138,7 @@ fn instructions<'input, E: ParseError<&'input [u8]>>(
 
         0x01 => {
             consume!((input, argument_0) = leb(input)?);
-            (input, Instruction::Call(argument_0))
+            (input, Instruction::Call(argument_0 as usize))
         }
 
         0x02 => {
